auth-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
pydantic-settings
passlib
python-jose
python-jose[cryptography]
google-auth
google-auth-oauthlib
python-multipart